requirements.txt
-i https://pypi.org/simple
arrow==0.15.7
asgiref==3.2.10
coloredlogs==14.0
django-cors-headers==3.4.0
django-environ==0.4.5
django==3.0.8
djangorestframework-jwt==1.11.0
djangorestframework==3.11.0
humanfriendly==8.2
mysqlclient==2.0.1
pillow==7.2.0
pyjwt==1.7.1
pymysql==0.9.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
gunicorn==20.0.4
